refactor: player 관련 로직 분리 (#121)
